+++ build.gradle
@@ -1,95 +1,56 @@
 plugins {
-	id 'fabric-loom' version '0.6-SNAPSHOT'
-	id 'maven-publish'
+    id "architectury-plugin" version "3.4-SNAPSHOT"
+    id "dev.architectury.loom" version "0.11.0-SNAPSHOT" apply false
+    id 'maven-publish'
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-archivesBaseName = project.archives_base_name
-version = project.mod_version
-group = project.maven_group
-
-repositories {
-	mavenLocal()
-	mavenCentral()
-	maven {
-		url 'http://maven.fabricmc.net/'
-		name 'Fabric'
-	}
-	maven {
-		name = 'JitPack'
-		url = 'https://jitpack.io'
-	}
-
-	jcenter()
+architectury {
+    minecraft = rootProject.minecraft_version
 }
 
-dependencies {
-	minecraft "com.mojang:minecraft:${project.minecraft_version}"
-	mappings "net.fabricmc:yarn:${project.yarn_mappings}:v2"
-	modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
-	compileOnly "com.google.code.findbugs:jsr305:3.0.2"
+subprojects {
+    apply plugin: "dev.architectury.loom"
 
-	modCompileOnly "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
-	modRuntime "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
+    loom {
+        silentMojangMappingsLicense()
+    }
 
-	modCompileOnly "com.github.CaffeineMC.lithium-fabric:lithium:mc1.16.5-0.6.4"
-	modCompileOnly "com.github.CaffeineMC.lithium-fabric:lithium-api:mc1.16.5-0.6.4"
-
-	modCompileOnly "com.github.qouteall.ImmersivePortalsMod:imm_ptl_core:v0.90-1.16"
+    dependencies {
+        minecraft "com.mojang:minecraft:${rootProject.minecraft_version}"
+        mappings "net.fabricmc:yarn:${rootProject.minecraft_version}+build.${project.yarn_mappings}:v2"
+        compileOnly "com.google.code.findbugs:jsr305:3.0.2"
+    }
 }
 
-processResources {
-	inputs.property "version", project.version
+allprojects {
+    apply plugin: "java"
+    apply plugin: "architectury-plugin"
+    apply plugin: "maven-publish"
 
-	from(sourceSets.main.resources.srcDirs) {
-		include "fabric.mod.json"
-		expand "version": project.version
-	}
+    archivesBaseName = rootProject.archives_base_name
+    version = rootProject.mod_version
+    group = rootProject.maven_group
 
-	from(sourceSets.main.resources.srcDirs) {
-		exclude "fabric.mod.json"
-	}
-}
+    repositories {
+        maven {
+            name = 'JitPack'
+            url = 'https://jitpack.io'
+        }
+    }
 
-// ensure that the encoding is set to UTF-8, no matter what the system default is
-// this fixes some edge cases with special characters not displaying correctly
-// see http://yodaconditions.net/blog/fix-for-java-file-encoding-problems-with-gradle.html
-tasks.withType(JavaCompile) {
-	options.encoding = "UTF-8"
-}
+    tasks.withType(JavaCompile) {
+        options.encoding = "UTF-8"
 
-// Loom will automatically attach sourcesJar to a RemapSourcesJar task and to the "build" task
-// if it is present.
-// If you remove this task, sources will not be generated.
-task sourcesJar(type: Jar, dependsOn: classes) {
-	classifier = "sources"
-	from sourceSets.main.allSource
-}
+        // The Minecraft launcher currently installs Java 8 for users, so your mod probably wants to target Java 8 too
+        // JDK 9 introduced a new way of specifying this that will make sure no newer classes or methods are used.
+        // We'll use that if it's available, but otherwise we'll use the older option.
+        def targetVersion = 8
+        if (JavaVersion.current().isJava9Compatible()) {
+            options.release = targetVersion
+        }
+    }
 
-jar {
-	from "LICENSE"
-}
-
-// configure the maven publication
-publishing {
-	publications {
-		mavenJava(MavenPublication) {
-			artifact(remapJar) {
-				builtBy remapJar
-			}
-			artifact(sourcesJar) {
-				builtBy remapSourcesJar
-			}
-		}
-	}
-	repositories {
-		maven {
-			url 's3://hephaestus.dev/release'
-			authentication {
-				awsIm(AwsImAuthentication)
-			}
-		}
-	}
+    java {
+        withSourcesJar()
+    }
 }
