+++ common/build.gradle
@@ -1,0 +1,27 @@
+dependencies {
+    // We depend on fabric loader here to use the fabric @Environment annotations and get the mixin dependencies
+    modImplementation "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
+    // Remove the next line if you don't want to depend on the API
+    modApi "me.shedaniel:architectury:${rootProject.architectury_version}"
+}
+
+architectury {
+    common()
+}
+
+configurations {
+    dev
+}
+
+artifacts {
+    dev(jar)
+}
+
+publishing {
+    publications {
+        mavenCommon(MavenPublication) {
+            artifactId = rootProject.archives_base_name
+            from components.java
+        }
+    }
+}
