+++ common/src/main/java/dev/hephaestus/fiblib/impl/FibLog.java
@@ -1,0 +1,29 @@
+package dev.hephaestus.fiblib.impl;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+@SuppressWarnings("unused")
+public class FibLog {
+	public static final String MOD_ID = "fiblib";
+	private static final String MOD_NAME = "FibLib";
+
+	private static final Logger LOGGER = LogManager.getLogger();
+    public static boolean DEBUG = false;
+
+	public static void log(String msg) {
+		log("%s", msg);
+	}
+
+	public static void log(String format, Object... args) {
+		LOGGER.info(String.format("[%s] %s", MOD_NAME, String.format(format, args)));
+	}
+
+	public static void debug(String msg) {
+		debug("%s", msg);
+	}
+
+	public static void debug(String format, Object... args) {
+		if (DEBUG) LOGGER.info(String.format("[%s] %s", MOD_NAME, String.format(format, args)));
+	}
+}
