+++ fabric/src/main/java/dev/hephaestus/fiblib/impl/FibLib.java
@@ -12,34 +12,12 @@ import net.minecraft.server.world.ThreadedAnvilChunkStorage;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.World;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 @SuppressWarnings("unused")
 public class FibLib {
-	public static final String MOD_ID = "fiblib";
-	private static final String MOD_NAME = "FibLib";
-
-	private static final Logger LOGGER = LogManager.getLogger();
     public static final boolean DEBUG = FabricLoader.getInstance().isDevelopmentEnvironment();
 
     public static final ThreadLocal<ServerPlayerEntity> PLAYER = new ThreadLocal<>();
-
-	public static void log(String msg) {
-		log("%s", msg);
-	}
-
-	public static void log(String format, Object... args) {
-		LOGGER.info(String.format("[%s] %s", MOD_NAME, String.format(format, args)));
-	}
-
-	public static void debug(String msg) {
-		debug("%s", msg);
-	}
-
-	public static void debug(String format, Object... args) {
-		if (DEBUG) LOGGER.info(String.format("[%s] %s", MOD_NAME, String.format(format, args)));
-	}
 
 	public static void resendChunks(ServerPlayerEntity player) {
 		ThreadedAnvilChunkStorage TACS = ((TACSAccessor) player.getServerWorld().getChunkManager())
