+++ fabric/src/main/java/dev/hephaestus/fiblib/mixin/VolatileMixinPlugin.java
