+++ fabric/src/main/java/dev/hephaestus/fiblib/mixin/packets/chunkdata/MixinArrayPalette.java
