+++ fabric/src/main/java/dev/hephaestus/fiblib/mixin/packets/chunkdata/MixinBiMapPalette.java
