+++ fabric/src/main/java/dev/hephaestus/fiblib/mixin/packets/chunkdata/MixinLithiumHashPalette.java
