+++ fabric/src/main/resources/fabric.mod.json
