+++ gradle.properties
@@ -1,12 +1,20 @@
-# Fabric Properties
-minecraft_version   = 1.16.5
-yarn_mappings       = 1.16.5+build.4
-loader_version      = 0.11.3
+org.gradle.jvmargs=-Xmx2048M
 
 # Mod Properties
-mod_version         = 1.0.2
-maven_group         = dev.hephaestus
-archives_base_name  = FibLib
+mod_version=1.0.3
+maven_group=dev.hephaestus
+archives_base_name=FibLib
 
-# Dependencies
-fabric_version      = 0.31.0+1.16
+# Shared Properties
+minecraft_version=1.16.5
+minecraft_major_version=1.16
+architectury_version=1.23.33
+yarn_mappings=10
+
+# Fabric Properties
+fabric_loader_version=0.13.3
+fabric_api_version=0.42.0
+lithium_version=0.6.4
+
+# Forge Properties
+forge_version=36.1.18
