+++ settings.gradle
@@ -1,10 +1,21 @@
 pluginManagement {
     repositories {
-        jcenter()
         maven {
-            name = 'Fabric'
-            url = 'https://maven.fabricmc.net/'
+            name = "Fabric"
+            url = "https://maven.fabricmc.net/"
+        }
+        maven {
+            name = "Architectury"
+            url = "https://maven.architectury.dev/"
+        }
+        maven {
+            name = "Forge"
+            url = "https://maven.minecraftforge.net/"
         }
         gradlePluginPortal()
     }
 }
+
+include("common")
+include("fabric")
+
